requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
addict=2.4.0=pypi_0
aliyun-python-sdk-core=2.14.0=pypi_0
aliyun-python-sdk-kms=2.16.2=pypi_0
appdirs=1.4.4=pypi_0
argcomplete=3.1.6=pypi_0
asttokens=2.4.0=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py310h7f8727e_1002
bypy=1.8.2=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
chumpy=0.70=pypi_0
click=8.1.7=pypi_0
clik=0.92.4=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.0.5=py310hdb19cb5_0
crcmod=1.7=pypi_0
cryptography=41.0.3=py310h130f0dd_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h9c0eb46_1
cython=3.0.6=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
decord=0.6.0=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.7=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
dv-processing=1.7.9=pypi_0
easygui=0.98.3=pypi_0
einops=0.7.0=pypi_0
engineering-notation=0.10.0=pypi_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=h6a678d5_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
flow-vis-torch=0.1=pypi_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.46.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.12.1=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.25.2=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
grpcio=1.60.0=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.10.0=pypi_0
huggingface-hub=0.19.4=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py310h06a4308_0
imageio=2.33.0=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.25.2=pyh2140261_0
ipython=8.12.3=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.2.0=py310h06a4308_0
jpeg=9e=h5eee18b_1
json-tricks=3.17.3=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.3.1=pyhd8ed1ab_0
jupyter_core=4.12.0=py310hff52083_0
jupyterlab-pygments=0.3.0=pypi_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.20.1=h568e23c_1
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=ha637b67_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libpng=1.6.39=h5eee18b_0
libpq=12.15=h37d81fd_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libuv=1.44.2=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hcbfbd50_0
libxslt=1.1.37=h2085143_0
llvmlite=0.41.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
mako=1.2.4=pypi_0
markdown=3.5.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
mat73=0.62=pypi_0
matplotlib=3.7.2=py310h06a4308_0
matplotlib-base=3.7.2=py310h1128e8f_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mmaction2=1.2.0=dev_0
mmcv=2.1.0=pypi_0
mmcv-full=1.7.1=pypi_0
mmdet=3.2.0=pypi_0
mmengine=0.10.1=pypi_0
mmpose=1.2.0=pypi_0
model-index=0.1.11=pypi_0
multiprocess=0.70.15=pypi_0
munkres=1.1.4=py_0
mysql=5.7.24=he378463_2
nbclient=0.9.0=pypi_0
nbconvert=7.16.1=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=pypi_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numba=0.58.1=pypi_0
numpy=1.26.0=py310h5f9d8c6_0
numpy-base=1.26.0=py310hb5e798b_0
oauthlib=3.2.2=pypi_0
opencv-contrib-python=4.8.1.78=pypi_0
opencv-python=4.8.1.78=pypi_0
opendatalab=0.0.10=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openmim=0.3.9=pypi_0
openssl=1.1.1w=h7f8727e_0
openxlab=0.0.31=pypi_0
ordered-set=4.1.0=pypi_0
oss2=2.17.0=pypi_0
packaging=23.1=py310h06a4308_0
pandas=2.1.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=23.2.1=py310h06a4308_0
pipreqs=0.5.0=pypi_0
platformdirs=3.11.0=pypi_0
ply=3.11=py310h06a4308_0
plyer=2.1.0=pypi_0
portalocker=2.8.2=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.23.4=pypi_0
psutil=5.9.0=py310h5eee18b_0
ptflops=0.7.1.2=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.19.0=pypi_0
pycuda=2022.2.2=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=23.2.0=py310h06a4308_0
pyparsing=3.0.9=py310h06a4308_0
pypushdeer=0.0.3=pypi_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.8=h7a1cb2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytools=2023.1.1=pypi_0
pytorch=1.11.0=py3.10_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytorch3d=0.7.5=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py310h6a678d5_0
qt-main=5.15.2=h5b8104b_9
qt-webengine=5.15.9=h9ab4d14_7
qtwebkit=5.212=h3fafdc1_5
readline=8.2=h5eee18b_0
referencing=0.33.0=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.4.2=pypi_0
rpds-py=0.18.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.1=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.3.0=py310h1128e8f_0
scipy=1.11.3=py310h5f9d8c6_0
screeninfo=0.8.1=pypi_0
sentry-sdk=1.39.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=60.2.0=pypi_0
shapely=2.0.2=pypi_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
soupsieve=2.5=pypi_0
sparseconvnet=0.2=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.4.0=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.9.26=pypi_0
timm=0.9.12=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=pypi_0
torchaudio=0.11.0=py310_cu113
torchvision=0.12.0=py310_cu113
tornado=6.3.3=py310h5eee18b_0
tqdm=4.65.2=pypi_0
traitlets=5.11.2=pyhd8ed1ab_0
typing_extensions=4.7.1=py310h06a4308_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=py310h06a4308_0
v2e=1.5.1=pypi_0
wandb=0.16.1=pypi_0
wcwidth=0.2.8=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
werkzeug=3.0.1=pypi_0
wget=3.2=pypi_0
wheel=0.41.2=py310h06a4308_0
xlsxwriter=3.1.1=py310h06a4308_0
xtcocotools=1.14.3=pypi_0
xz=5.4.2=h5eee18b_0
yacs=0.1.8=pypi_0
yapf=0.40.2=pypi_0
yarg=0.1.9=pypi_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
